rename module 'accounts' to 'addresses'
